Rename `parse_errors.py` to `parser.py`
I think it makes more sense to not name files based on actions,
but rather what the components of the file do. In this case, the
file consists of all the "parser" stuff.

The `run_and_get_errors()` function has also been moved out of
here since it is not directly related to parsing. It will be added
back in the next commit.