Added intermediary `get_all_error_results()` function
Since we can't pass any information into the Display class, but for
debugging purposes we also want to be able to test the program
without opening up the display, this commit adds the above function
where all the core logic has been added (and taken out of the
Display class)

Additionally, the `main()` function now checks for a `-n` flag, and
if it exists, the display is never started and the function is called
directly so any printed output can be seen.

If the flag does not exist, the Display is started. Internally, the
Display class also calls this function.