Moved argument parsing to `pytui.arguments`
Since we want to be able to access the command line arguments from
within methods called by the Display class (and we cannot pass in
any arguments), we need to use a global variable inside this file.

The arguments can be accessed by importing `pytui.arguments.args`,
which is a defaultdict, and returns None if the argument doesn't
exist.