Update display to use Markdown rendering
This patch takes the StackOverflow answer bodies and converts them
to markdown from HTML using the `html2text` library. Then, this
markdown is passed through the `Markdown` generator in rich which
allows us to get nicely formatted output.

For now, only the answers for the first question in the list of
results is shown. In the future, we need to:

  1. Also display the question on top
  2. Figure out how to let the user change between the different
     questions that were a part of the search result.

I've attempted to try to do (2) currently, but it seems there's
no built in list widget (where we can assign callbacks to each
button) - at least that I can see. The framework is still very
much incomplete. Chances are we may not be able to do something
much more sophisticated than what is currently being done.