Add syntax highlighting to inline/block code
Rich already supports markdown syntax highlighting, but
`markdownify` did not add the language specifier for the code
blocks when converting the HTML. Assuming (for now) that we are
always dealing with Python code, I created a custom subclass of
`MarkdownConverter` that adds the 'py' specifier for all the
markdown code blocks, which gives us pretty syntax-highlighted
code.

Also, I enabled syntax highlighting for inline code. This was just
a case of passing in the correct option into the `Markdown`
initializer.