Use the rich pretty renderer for tracebacks
Instead of just printing out the text, `rich` provides nice
pretty-printers to render out the exceptions in a much nicer to
see way with context of what's happening where. This patch uses
this inside the app when we display an exception.

Perhaps we also want to do this in the command line when we print
out an exception?