Unescape characters in the sidebar titles
Previously the sidebar question titles has html-escaped characters
such as `&#39`, which we weren't doing anything to sanitize. These
are not being unescaped with the built in `html` module.

It seems that we don't need to do this for the actual questions/
answers view since it's taken care of by the `Markdown()` code
(I think so at least, since the same issue doesn't show up there).