Remove extra frames from the traceback
Because of the way we are currently running the code, any traceback
created during the execution of the application will be include the
stack frames of `wtpython` which includes the functions within
`runpy` that eventually execute the code. However, showing this as
part of the traceback to the user would just be confusing to them
since the issue is likely not within our code.

This commit adds a function that creates a new traceback object from
the old one, but only keeping the frames which are actually relevant
to the user's code.